backend/requirements.txt
# Core backend packages
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
google-generativeai
pypdf
python-docx
openai-whisper
gTTS
Pillow
fpdf
requests
supabase
plotly
matplotlib
seaborn
nltk
sumy